requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
Marshmallow-SQLAlchemy
Flask-Restplus
pymysql
Flask-Migrate
Flask-Script